+++ src/pytest_recording/_vcr.py
@@ -35,15 +35,21 @@ class CombinedPersister(FilesystemPersister):
         #    but the functional approach is faster
         # 2. It could be done more efficient, but the `deserialize` implementation should be adjusted as well
         #    But it is a private API, which could be changed.
-        return starmap(unpack, zip(*all_content))
+        requests, responses = starmap(unpack, zip(*all_content))
+        requests, responses = list(requests), list(responses)
+        if not requests or not responses:
+            raise ValueError("No cassettes found.")
+        return requests, responses
 
 
 # pylint: disable=too-many-arguments
 def use_cassette(default_cassette, vcr_cassette_dir, record_mode, markers, config, pytestconfig):
     """Create a VCR instance and return an appropriate context manager for the given cassette configuration."""
     merged_config = merge_kwargs(config, markers)
+    if "record_mode" in merged_config:
+        record_mode = merged_config["record_mode"]
     path_transformer = get_path_transformer(merged_config)
-    if record_mode == "rewrite" or merged_config.get("record_mode") == "rewrite":
+    if record_mode == "rewrite":
         path = path_transformer(os.path.join(vcr_cassette_dir, default_cassette))
         try:
             os.remove(path)